requirements.txt
fastapi
uvicorn
requests
xmltodict
python-magic-bin; platform_system == "Windows"
python-magic; platform_system == "Linux"
https://github.com/hajle-silesia/common/zipball/master
kafka-python
